Rename to ExtractRetryDelay and add an example for
NewReadWriteStmtBasedTransaction.
