+++ spanner/client_test.go
@@ -591,96 +591,6 @@ func testReadOnlyTransaction(t *testing.T, executionTimes map[string]SimulatedEx
 	return executeSingerQuery(ctx, tx)
 }
 
-func TestClient_ReadWriteTransaction_StatementBased(t *testing.T) {
-	t.Parallel()
-
-	rowCount, attempts, err := testReadWriteStmtBasedTransaction(t, make(map[string]SimulatedExecutionTime))
-	if err != nil {
-		t.Fatalf("transaction failed to commit: %v", err)
-	}
-	if rowCount != SelectSingerIDAlbumIDAlbumTitleFromAlbumsRowCount {
-		t.Fatalf("Row count mismatch, got %v, expected %v", rowCount, SelectSingerIDAlbumIDAlbumTitleFromAlbumsRowCount)
-	}
-	if g, w := attempts, 1; g != w {
-		t.Fatalf("number of attempts mismatch:\nGot%d\nWant:%d", g, w)
-	}
-}
-
-func TestClient_ReadWriteTransaction_StatementBased_CommitAborted(t *testing.T) {
-	t.Parallel()
-	rowCount, attempts, err := testReadWriteStmtBasedTransaction(t, map[string]SimulatedExecutionTime{
-		MethodCommitTransaction: {Errors: []error{status.Error(codes.Aborted, "Transaction aborted")}},
-	})
-	if err != nil {
-		t.Fatalf("transaction failed to commit: %v", err)
-	}
-	if rowCount != SelectSingerIDAlbumIDAlbumTitleFromAlbumsRowCount {
-		t.Fatalf("Row count mismatch, got %v, expected %v", rowCount, SelectSingerIDAlbumIDAlbumTitleFromAlbumsRowCount)
-	}
-	if g, w := attempts, 2; g != w {
-		t.Fatalf("number of attempts mismatch:\nGot%d\nWant:%d", g, w)
-	}
-}
-
-func testReadWriteStmtBasedTransaction(t *testing.T, executionTimes map[string]SimulatedExecutionTime) (rowCount int64, attempts int, err error) {
-	server, client, teardown := setupMockedTestServer(t)
-	defer teardown()
-	for method, exec := range executionTimes {
-		server.TestSpanner.PutExecutionTime(method, exec)
-	}
-	ctx := context.Background()
-
-	f := func(tx *ReadWriteStmtBasedTransaction) (int64, error) {
-		iter := tx.Query(ctx, NewStatement(SelectSingerIDAlbumIDAlbumTitleFromAlbums))
-		defer iter.Stop()
-		rowCount := int64(0)
-		for {
-			row, err := iter.Next()
-			if err == iterator.Done {
-				break
-			}
-			if err != nil {
-				return 0, err
-			}
-			var singerID, albumID int64
-			var albumTitle string
-			if err := row.Columns(&singerID, &albumID, &albumTitle); err != nil {
-				return 0, err
-			}
-			rowCount++
-		}
-		return rowCount, nil
-	}
-
-	for {
-		attempts++
-		tx, err := NewReadWriteStmtBasedTransaction(ctx, client)
-		if err != nil {
-			return 0, attempts, fmt.Errorf("failed to begin a transaction: %v", err)
-		}
-		rowCount, err = f(tx)
-		if err != nil && status.Code(err) != codes.Aborted {
-			tx.Rollback(ctx)
-			return 0, attempts, err
-		} else if err == nil {
-			_, err = tx.Commit(ctx)
-			if err == nil {
-				break
-			} else if status.Code(err) != codes.Aborted {
-				return 0, attempts, err
-			}
-		}
-		// Set a default sleep time if the server delay is absent.
-		delay := 10 * time.Millisecond
-		if serverDelay, hasServerDelay := extractRetryDelay(err); hasServerDelay {
-			delay = serverDelay
-		}
-		time.Sleep(delay)
-	}
-
-	return rowCount, attempts, err
-}
-
 func TestClient_ReadWriteTransaction(t *testing.T) {
 	t.Parallel()
 	if err := testReadWriteTransaction(t, make(map[string]SimulatedExecutionTime), 1); err != nil {
