+++ spanner/examples_test.go
@@ -26,6 +26,8 @@ import (
 	"cloud.google.com/go/spanner"
 	"google.golang.org/api/iterator"
 	sppb "google.golang.org/genproto/googleapis/spanner/v1"
+	"google.golang.org/grpc/codes"
+	"google.golang.org/grpc/status"
 )
 
 func ExampleNewClient() {
@@ -709,4 +711,59 @@ func ExampleKeySets() {
 
 	_ = ks //TODO: Go use the KeySet in another query.
 
+}
+
+func ExampleNewReadWriteStmtBasedTransaction() {
+	ctx := context.Background()
+	client, err := spanner.NewClient(ctx, myDB)
+	if err != nil {
+		// TODO: Handle error.
+	}
+	defer client.Close()
+
+	f := func(tx *spanner.ReadWriteStmtBasedTransaction) error {
+		var balance int64
+		row, err := tx.ReadRow(ctx, "Accounts", spanner.Key{"alice"}, []string{"balance"})
+		if err != nil {
+			return err
+		}
+		if err := row.Column(0, &balance); err != nil {
+			return err
+		}
+
+		if balance <= 10 {
+			return errors.New("insufficient funds in account")
+		}
+		balance -= 10
+		m := spanner.Update("Accounts", []string{"user", "balance"}, []interface{}{"alice", balance})
+		return tx.BufferWrite([]*spanner.Mutation{m})
+	}
+
+	for {
+		tx, err := spanner.NewReadWriteStmtBasedTransaction(ctx, client)
+		if err != nil {
+			// TODO: Handle error.
+			break
+		}
+		err = f(tx)
+		if err != nil && status.Code(err) != codes.Aborted {
+			tx.Rollback(ctx)
+			// TODO: Handle error.
+			break
+		} else if err == nil {
+			_, err = tx.Commit(ctx)
+			if err == nil {
+				break
+			} else if status.Code(err) != codes.Aborted {
+				// TODO: Handle error.
+				break
+			}
+		}
+		// Set a default sleep time if the server delay is absent.
+		delay := 10 * time.Millisecond
+		if serverDelay, hasServerDelay := spanner.ExtractRetryDelay(err); hasServerDelay {
+			delay = serverDelay
+		}
+		time.Sleep(delay)
+	}
 }
